requirements.txt
# Web Framework
Flask==3.0.0
gunicorn==21.2.0

# AI/ML Libraries
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# OCR & Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.0

# NLP & Medical Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Utilities
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# Production Dependencies
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0

# For model loading and caching
joblib==1.3.2
tqdm==4.66.1
filelock==3.13.1